+++ corebehrt/constants/train.py
@@ -1,0 +1,2 @@
+DEFAULT_CV_SPLITS = 5
+DEFAULT_VAL_SPLIT = 0.2
